Don't complain about missing DISTRO or MACHINE if local.conf exists
If local.conf exists, assume it contains correct values for
DISTRO and MACHINE, if they have not been provided in the
environment.  If DISTRO and/or MACHINE have been provided in the
environment and are valid, use the provided ones.

If local.conf doesn't exist, require valid values for both DISTRO
and MACHINE in the environment.

Signed-off-by: Mario Domenech Goulart <mario@ossystems.com.br>